+++ models/ActiveRecord.php
@@ -6,7 +6,7 @@ namespace wdmg\base\models;
  * Yii2 ActiveRecord
  *
  * @category        Model
- * @version         1.2.1
+ * @version         1.2.2
  * @author          Alexsander Vyshnyvetskyy <alex.vyshnyvetskyy@gmail.com>
  * @link            https://github.com/wdmg/yii2-base
  * @copyright       Copyright (c) 2019 - 2020 W.D.M.Group, Ukraine